requirements.txt
streamlit==1.54.0
torch==2.11.0
torchvision==0.26.0
Pillow==12.1.0
